Drag and drop file opening, open files by association, show returned value and other fixes

+++ Xtv.Common/Abstractions/ITrace.cs
@@ -14,7 +14,7 @@ namespace Xtv.Common.Abstractions
     public interface ITrace
     {
         byte Level { get; }
-        short CallIndex { get; }
+        ulong CallIndex { get; }
         TraceCall Call { get; }
         bool IsUserDefined { get; }
         string[] Parameters { get; }
